refactor: Extract formatRelativeTime to shared utility
- Create shared utility function in src/lib/utils/date.ts
- Remove duplicated formatRelativeTime from 3 components
- Update imports to use shared utility

Related to #17
