+++ src/app/dashboard/reports/page.tsx
@@ -10,6 +10,7 @@ import { ReportStatusBadge } from '@/features/dashboard/components/ReportStatusB
 import type { IReportStatus } from '@/features/dashboard/types';
 import { Input } from '@/components/ui/input';
 import { Skeleton } from '@/components/ui/skeleton';
+import { formatRelativeTime } from '@/lib/utils/date';
 import {
   Select,
   SelectContent,
@@ -78,21 +79,6 @@ export default function ReportsListPage() {
     observer.observe(sentinel);
     return () => observer.disconnect();
   }, [handleIntersect]);
-
-  const formatRelativeTime = (dateStr: string) => {
-    const date = new Date(dateStr);
-    const now = new Date();
-    const diffMs = now.getTime() - date.getTime();
-    const diffMins = Math.floor(diffMs / 60000);
-    const diffHours = Math.floor(diffMs / 3600000);
-    const diffDays = Math.floor(diffMs / 86400000);
-
-    if (diffMins < 1) return 'Baru saja';
-    if (diffMins < 60) return `${diffMins} menit lalu`;
-    if (diffHours < 24) return `${diffHours} jam lalu`;
-    if (diffDays < 7) return `${diffDays} hari lalu`;
-    return date.toLocaleDateString('id-ID', { day: 'numeric', month: 'short', year: 'numeric' });
-  };
 
   const handleSortChange = (value: string) => {
     const [field, dir] = value.split(':');
